Added book class and edit button
Added book class, which has a constructor of title, author, pages and is
read. Also Added edit button, which lets the user to edit the title,
author and pages.

The edit button firstly takes gets the values from myLibrary array and
puts them to edit form. I achieved this by setting the currentEditIndex
to keep track of index of the book being edited. Then the data is
returned and assigned to form. Then after the user edits the data and
the submit button is pressed the values are assigned to the array using
the currentEditIndex and the book card is updated.